refactor: ScheduleDropdown 및 Schedule 분리 [TRL-199]
